往Compose迁移中，主页面已经可以使用[20210929]
Former-commit-id: b2dfa4a8ddabe3e13e5ff7809c19e56c90aea17e

+++ codeeditor/src/main/java/com/funny/translation/codeeditor/ui/runner/CodeRunnerViewModel.kt
@@ -29,8 +29,8 @@ class CodeRunnerViewModel(application: Application) : BaseViewModel(application)
                 onSuccess = {
                     val jsTranslateTask = JsTranslateTask(
                         jsEngine = this,
-                        sourceLanguage = activityCodeViewModel.sourceLanguage.value!!.toShort(),
-                        targetLanguage = activityCodeViewModel.targetLanguage.value!!.toShort(),
+                        sourceLanguage = activityCodeViewModel.sourceLanguage.value!!,
+                        targetLanguage = activityCodeViewModel.targetLanguage.value!!,
                         sourceString = activityCodeViewModel.sourceString.value!!
                     )
                     Coroutine.async {
